chore: fix & reorganize broadcaster entity structure
