+++ persistence-spring-jpa/src/main/java/io/naryo/infrastructure/configuration/persistence/entity/broadcaster/configuration/BroadcasterCacheEntity.java
@@ -6,20 +6,16 @@ import io.naryo.application.configuration.source.model.broadcaster.configuration
 import jakarta.persistence.Embeddable;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Embeddable
 @Getter
+@NoArgsConstructor
 public class BroadcasterCacheEntity implements BroadcasterCacheConfigurationDescriptor {
-
-    private static final Duration DEFAULT_EXPIRATION_TIME = Duration.ofMinutes(5);
 
     private @NotNull Duration expirationTime;
 
     public BroadcasterCacheEntity(Duration expirationTime) {
-        this.expirationTime = (expirationTime != null) ? expirationTime : DEFAULT_EXPIRATION_TIME;
-    }
-
-    public BroadcasterCacheEntity() {
-        this(DEFAULT_EXPIRATION_TIME);
+        this.expirationTime = expirationTime;
     }
 }
