+++ persistence-spring-jpa/src/main/java/io/naryo/infrastructure/configuration/persistence/repository/broadcaster/BroadcasterConfigurationEntityRepository.java
@@ -1,7 +1,9 @@
 package io.naryo.infrastructure.configuration.persistence.repository.broadcaster;
 
+import java.util.UUID;
+
 import io.naryo.infrastructure.configuration.persistence.entity.broadcaster.configuration.BroadcasterConfigurationEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BroadcasterConfigurationEntityRepository
-        extends JpaRepository<BroadcasterConfigurationEntity, String> {}
+        extends JpaRepository<BroadcasterConfigurationEntity, UUID> {}
