+++ persistence-spring-jpa/src/main/java/io/naryo/infrastructure/configuration/persistence/repository/broadcaster/BroadcasterEntityRepository.java
@@ -1,6 +1,8 @@
 package io.naryo.infrastructure.configuration.persistence.repository.broadcaster;
 
-import io.naryo.infrastructure.configuration.persistence.entity.broadcaster.target.BroadcasterEntity;
+import java.util.UUID;
+
+import io.naryo.infrastructure.configuration.persistence.entity.broadcaster.BroadcasterEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BroadcasterEntityRepository extends JpaRepository<BroadcasterEntity, String> {}
+public interface BroadcasterEntityRepository extends JpaRepository<BroadcasterEntity, UUID> {}
