+++ persistence-spring-jpa/src/main/java/io/naryo/infrastructure/configuration/provider/JpaBroadcasterSourceProvider.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 import io.naryo.application.configuration.source.model.broadcaster.BroadcasterDescriptor;
 import io.naryo.application.configuration.source.provider.broadcaster.BroadcasterSourceProvider;
-import io.naryo.infrastructure.configuration.persistence.entity.broadcaster.target.BroadcasterEntity;
+import io.naryo.infrastructure.configuration.persistence.entity.broadcaster.BroadcasterEntity;
 import io.naryo.infrastructure.configuration.persistence.repository.broadcaster.BroadcasterEntityRepository;
 import org.springframework.stereotype.Component;
